Build the component selection reason when required
We were building the component selection reason incrementally, as a
`ComponentState` was chosen for a `SelectorState`. This meant that
reasons could have ended up on evicted nodes, and were missing from the
actual selected node.

Instead, we now build the reason for a component based on the set of
selectors that were involved in choosing that component, even if those
selectors were upgraded through conflict resolution.